re-enable epel repository
Since Tendrl itself finally starts to cover package installation to some
degree, we need to have epel enabled from now on for this to work.

Note: when you run these playbooks on RHEL, the setup fails - there is
no change in the check. This is a feature: enabling EPEL in such case
is a blocker bug and we need to catch it as soon as possible.